requirements.txt
channels==2.2.0
Django==2.2.3
django-baton==1.5.4
django-debug-toolbar==2.0
django-money==0.15
django-safedelete==0.5.1
Pillow==6.1.0
psycopg2==2.8.3
pytest-asyncio==0.10.0
pytest-django==3.5.1
